# openapi.json
{
    "openapi": "3.0.0",
    "info": {
      "title": "Task API",
      "version": "1.0.0",
      "description": "A simple in-memory CRUD API for managing tasks"
    },
    "servers": [{ "url": "http://localhost:3000" }],
    "paths": {
      "/tasks": {
        "get": {
          "summary": "Get all tasks",
          "responses": { "200": { "description": "List of tasks" } }
        },
        "post": {
          "summary": "Create a new task",
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": { "title": { "type": "string" } },
                  "required": ["title"]
                }
              }
            }
          },
          "responses": {
            "201": { "description": "Task created" },
            "400": { "description": "Invalid input" }
          }
        }
      },
      "/tasks/{id}": {
        "get": {
          "summary": "Get a single task",
          "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
          "responses": {
            "200": { "description": "Task found" },
            "404": { "description": "Task not found" }
          }
        },
        "put": {
          "summary": "Update a task",
          "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
          "requestBody": {
            "required": true,
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "properties": {
                    "title": { "type": "string" },
                    "done": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "responses": {
            "200": { "description": "Task updated" },
            "400": { "description": "Invalid input" },
            "404": { "description": "Task not found" }
          }
        },
        "delete": {
          "summary": "Delete a task",
          "parameters": [{ "name": "id", "in": "path", "required": true, "schema": { "type": "integer" } }],
          "responses": {
            "204": { "description": "Task deleted" },
            "404": { "description": "Task not found" }
          }
        }
      }
    }
  }